refactor(popover): move styles to another file and use Text from Yoga
